+++ src/trx/game/effects/draw.c
@@ -25,7 +25,7 @@ void Effect_Draw(const int16_t effect_num)
         Output_DrawSprite(
             effect->interp.result.pos.x, effect->interp.result.pos.y,
             effect->interp.result.pos.z, Object_Get(O_GLOW)->mesh_idx,
-            effect->shade, COLOR_RGB_F_WHITE, DRAW_BLEND);
+            effect->shade, COLOR_RGB_F_WHITE, DRAW_BLEND, 1.0f);
         return;
     }
 
@@ -48,7 +48,7 @@ void Effect_Draw(const int16_t effect_num)
         Output_DrawSprite(
             effect->interp.result.pos.x, effect->interp.result.pos.y,
             effect->interp.result.pos.z, obj->mesh_idx - effect->frame_num,
-            shade, tint, DRAW_BLEND);
+            shade, tint, DRAW_BLEND, 1.0f);
     } else {
         Matrix_Push();
         Matrix_TranslateAbs32(effect->interp.result.pos);